pyproject.toml
[project]
name = "ai-calendar"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi[standard]>=0.128.4",
    "huggingface-hub>=1.4.1",
    "mcp>=1.26.0",
    "ollama>=0.6.1",
    "python-dotenv>=1.2.1",
]
